feat(theme): taking the lazy way out and committing the whole theme
I'm sorry, Mama. I never meant to hurt you.

The new theme has been tested back to IE11, and in <= IE10 the styles aren't rendered at all (plain-text version of the site) to make the site usable-if-not-fucking-ugly for people on older browsers.

I should have been committing these features in one at a time, but I chose to be a terrible person, ignore that part of my brain that makes good decisions, and ultimately to cop out with this übercommit.