draft(panic): light edits and notes on resolving this post
I’m still a little nervous that this sounds too whiny and over-
privileged, but after a conversation with Ivan, I get the sense
that — whether or not this problem is a “real” problem — it’s not
uncommon. And maybe it’s time for me to put something more
vulnerable than “here’s this cute story about how I made a totally
non-threatening mistake and then found a way to spin it to my
advantage” out there.

Or maybe I’m just a whiny asshole. Either way. I think this one’s
actually going to go live at some point.

re #10