wip: search is working but doesn’t look great
- posts are sent to Algolia for indexing
- search is available for title, description, body text, and SEO title

Right now the description always shows, even if the text is what actually matches. This is because the text is uploaded as raw Markdown. I don’t know if I really want to get into parsing Markdown to show the actual text match (footnotes will be a nightmare, for example), so maybe this just is what it is?

- [ ] Set the `BUILD_ALGOLIA_INDEX` env var for `master` builds on Netlify
- [ ] Style the search button in the nav
- [ ] Add controls for closing the search overlay
- [ ] Make it more obvious how searching works
- [ ] Generally make things look less like shit

re #56